requirements.txt
easygui
SpeechRecognition
pillow
numpy
PyAudio